+++ src/share/vector_index/ob_plugin_vector_index_adaptor.h
@@ -242,9 +242,9 @@ public:
   {}
 
   std::string Name() override { return "ObVsagSearchAlloc"; }
-  void* Allocate(size_t size) override;
+  void* Allocate(uint64_t size) override;
   void Deallocate(void* p) override { alloc_.free(p); };
-  void* Reallocate(void* p, size_t size) override;
+  void* Reallocate(void* p, uint64_t size) override;
   int64_t hold() { return alloc_.total(); }
   int64_t used() { return alloc_.used(); }
   void reset() { alloc_.reset(); }
